requirements.txt
dash==2.7.0
kafka-python==2.0.2
redis==4.3.4
